# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Rezervacija resursa",
        "description": "API za login, registraciju i upravljanje korisnicima",
        "termsOfService": "",
        "contact": {
            "email": "mislav.zabcic@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "admin/items/details/create/devicetype": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Kreiranje device tipa",
                "description": "Kreiranje device tipa",
                "operationId": "api.admin.items.details.create.devicetype",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "label",
                        "in": "body",
                        "description": "Labela",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "body",
                        "description": "Opis",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device type created"
                    },
                    "409": {
                        "description": "Item with that label already exists",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "403": {
                        "description": "No admin rights",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
        "admin/items/details/create/subtype": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Kreiranje  podtipa",
                "description": "Kreiranje podtipa",
                "operationId": "api.admin.items.details.create.subtype",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "label",
                        "in": "body",
                        "description": "Labela",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "body",
                        "description": "Opis",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subtype created"
                    },
                    "409": {
                        "description": "Item with that label already exists",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "403": {
                        "description": "No admin rights",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
        "reservations/request": {
            "post": {
                "tags": [
                    "reservations"
                ],
                "summary": "Zatra\u017ei opremu",
                "description": "Zatra\u017ei opremu",
                "operationId": "api.reservations.request",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "body",
                        "description": "Datum po\u010detka posudbe",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "return_date",
                        "in": "body",
                        "description": "Datum vra\u0107anja opreme, null ukoliko ne zna\u0161 kolko ti treba",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "item_id",
                        "in": "body",
                        "description": "Ime",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "type": "string"
                    },
                    {
                        "name": "remark",
                        "in": "body",
                        "description": "Napomena uz posudbu",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request made"
                    },
                    "409": {
                        "description": "User exists",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid form data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
        "reservations": {
            "get": {
                "tags": [
                    "reservations"
                ],
                "summary": "Dohvati sve rezervacije",
                "description": "Dohvati sve items ili filtrirane",
                "operationId": "api.reservations",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Pisati u formatu <ime_kolone_u_tablici>=<pojam_za_pretra\u017eivanje>",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Item"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "501": {
                        "description": "Invalid search data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
        "reservations/delete/{id}": {
            "delete": {
                "tags": [
                    "reservations"
                ],
                "summary": "Obri\u0161i korisnika",
                "description": "Obri\u0161i rezervaciju sa danim ID-om",
                "operationId": "api.reservations.delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "JWT token",
                        "required": true,
                        "type": "string",
                        "items": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id rezervacije",
                        "required": true,
                        "type": "integer",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Korisnik je obrisan"
                    },
                    "404": {
                        "description": "No data found",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "401": {
                        "description": "Token invalid",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "402": {
                        "description": "No token recived",
                        "schema": {
                            "$ref": "#/definitions/CustomError"
                        }
                    },
                    "410": {
                        "description": "Token expired",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    },
                    "403": {
                        "description": "Not your reservation",
                        "schema": {
                            "$ref": "#/definitions/TokenExpired"
                        }
                    }
                },
                "schemes": [
                    "http"
                ]
            }
        },
...
    },
    "definitions": {
        "CustomError": {
            "properties": {
                "error": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                }
            },
            "type": "object"
        },
...
        "DeviceType": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "label": {
                    "description": "",
                    "type": "integer",
                    "format": "string"
                },
                "description": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                }
            },
            "type": "object"
        },
        "Item": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "identifier": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "description": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "picture": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "kit_id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "type_id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "subtype_id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "device_type_id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "file": {
                    "description": "",
                    "format": "file"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "kit": {
                    "description": "",
                    "format": "Kit",
                    "$ref": "#/definitions/Kit"
                },
                "type": {
                    "description": "",
                    "format": "Type",
                    "$ref": "#/definitions/Type"
                },
                "subtype": {
                    "description": "",
                    "format": "Subtype",
                    "$ref": "#/definitions/SubType"
                },
                "device_type": {
                    "description": "",
                    "format": "DeviceType",
                    "$ref": "#/definitions/DeviceType"
                }
            },
            "type": "object"
        },
        "Kit": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "name": {
                    "description": "",
                    "type": "integer",
                    "format": "string"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                }
            },
            "type": "object"
        },
...
        "SubType": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "label": {
                    "description": "",
                    "type": "integer",
                    "format": "string"
                },
                "description": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                }
            },
            "type": "object"
        },
        "TokenExpired": {
            "properties": {
                "token": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "error": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                }
            },
            "type": "object"
        },
        "Type": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "format": "int"
                },
                "label": {
                    "description": "",
                    "type": "integer",
                    "format": "string"
                },
                "description": {
                    "description": "",
                    "type": "string",
                    "format": "string"
                },
                "created_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                },
                "updated_at": {
                    "description": "",
                    "type": "string",
                    "format": "date"
                }
            },
            "type": "object"
        },
...
    },
...
}